requirements.txt
absl-py==2.1.0
accelerate @ git+https://github.com/huggingface/accelerate.git@7aafa25673f4324cbcaeef508a13e09576f12dc4
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bert-score==0.3.13
bitsandbytes==0.42.0
captum==0.7.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
datasets==2.16.1
decorator==5.1.1
deep-translator==1.11.4
dill==0.3.7
docker-pycreds==0.4.0
docstring-parser==0.15
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.0
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
gradio==4.14.0
gradio_client==0.8.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.4
httpx==0.26.0
huggingface-hub==0.19.4
hyperframe==5.2.0
idna==2.10
importlib-resources==6.1.1
install==1.3.5
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
pandas==2.1.3
parso==0.8.3
peft @ git+https://github.com/huggingface/peft.git@189a9a666dc90f335b6a4db77f4b856b6a8bbdfd
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-multipart==0.0.6
python-telegram-bot==20.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rfc3986==1.5.0
rich==13.7.0
rouge-score==0.1.2
rpds-py==0.17.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentence-transformers==2.5.0
sentencepiece==0.1.99
sentry-sdk==1.38.0
setproctitle==1.3.3
shellingham==1.5.4
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.14.0
transformers @ git+https://github.com/huggingface/transformers@e694e985d770d14266dfd53378575a5d2a90256f
transformers-interpret==0.10.0
triton==2.1.0
trl==0.7.10
typer==0.9.0
typing_extensions==4.8.0
tyro==0.7.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
wandb==0.16.1
wcwidth==0.2.12
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.3

transformers~=4.38.0.dev0